# AstraServerGo.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "AstraServerGo",
    "description": "基于 `AstraScheduleServerGo` 源码生成的 OpenAPI 3.0 文档。\n\n- 共 55 个 API 端点，覆盖客户端课表、天气、WebSocket、管理后台认证、用户管理、学校/年级/班级管理、配置管理、自动任务、倒数日、备份导入导出等功能。\n- JWT 认证接口使用 `jwtAuth` 安全方案（通过 POST /web/auth/login 获取 Token）。\n- 写入接口同时支持 BasicAuth（兼容旧版客户端）和 JWT + 密码验证。\n- 部分接口返回结构是动态 JSON（`map[string]interface{}`），已尽量按实际行为建模。\n",
    "version": "1.1.0"
  },
...
  "paths": {
...
    "/web/autorun/compensation/holiday/{year}/{month}/{day}": {
      "get": {
        "summary": "根据节假日查询对应调休工作日",
        "operationId": "compensationFromHoliday",
        "tags": ["web-autorun"],
        "parameters": [
          {
            "name": "year", "in": "path", "description": "年份", "required": true,
            "schema": { "type": "integer" }, "example": 2026
          },
          {
            "name": "month", "in": "path", "description": "月份", "required": true,
            "schema": { "type": "integer", "minimum": 1, "maximum": 12 }, "example": 5
          },
          {
            "name": "day", "in": "path", "description": "日", "required": true,
            "schema": { "type": "integer", "minimum": 1, "maximum": 31 }, "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "查询结果",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CompensationSingleResponse" }
              }
            }
          },
          "400": {
            "description": "参数错误",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DetailResponse" }
              }
            }
          }
        },
        "security": []
      }
    },
    "/web/autorun/compensation/workday/{year}/{month}/{day}": {
      "get": {
        "summary": "根据工作日查询对应调休节假日",
        "operationId": "compensationFromWorkday",
        "tags": ["web-autorun"],
        "parameters": [
          {
            "name": "year", "in": "path", "description": "年份", "required": true,
            "schema": { "type": "integer" }, "example": 2026
          },
          {
            "name": "month", "in": "path", "description": "月份", "required": true,
            "schema": { "type": "integer", "minimum": 1, "maximum": 12 }, "example": 5
          },
          {
            "name": "day", "in": "path", "description": "日", "required": true,
            "schema": { "type": "integer", "minimum": 1, "maximum": 31 }, "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "查询结果",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CompensationSingleResponse" }
              }
            }
          },
          "400": {
            "description": "参数错误",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DetailResponse" }
              }
            }
          }
        },
        "security": []
      }
    },
    "/web/autorun/compensation/year/{year}": {
      "get": {
        "summary": "查询某年全部调休对照",
        "operationId": "compensationFromYear",
        "tags": ["web-autorun"],
        "parameters": [
          {
            "name": "year", "in": "path", "description": "年份", "required": true,
            "schema": { "type": "integer" }, "example": 2026
          }
        ],
        "responses": {
          "200": {
            "description": "查询结果",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CompensationYearResponse" }
              }
            }
          },
          "400": {
            "description": "参数错误",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DetailResponse" }
              }
            }
          }
        },
        "security": []
      }
    },
    "/web/schedule/by-date": {
      "get": {
        "summary": "按日期出课节列表",
        "description": "根据指定日期和 scope（school/grade/class），结合自动任务规则计算出当天的课节列表。",
        "operationId": "getScheduleByDate",
        "tags": ["web-schedule"],
        "parameters": [
          {
            "name": "date", "in": "query", "description": "日期，格式 YYYY-MM-DD", "required": true,
            "schema": { "type": "string", "format": "date" }
          },
          {
            "name": "scope", "in": "query", "description": "作用域，格式 school/grade/class", "required": true,
            "schema": { "type": "string" }, "example": "default/2024级/1班"
          }
        ],
        "responses": {
          "200": {
            "description": "课节列表",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["data"],
                  "properties": {
                    "data": {
                      "type": "object",
                      "required": ["periods"],
                      "properties": {
                        "periods": {
                          "type": "array",
                          "items": { "$ref": "#/components/schemas/Period" }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "参数错误",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DetailResponse" }
              }
            }
          }
        },
        "security": []
      }
    },
...
    "/web/auth/change-password": {
      "post": {
        "summary": "修改密码",
        "operationId": "changePassword",
        "tags": ["web-auth"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["old_password", "new_password"],
                "properties": {
                  "old_password": { "type": "string" },
                  "new_password": { "type": "string", "minLength": 6 },
                  "new_username": { "type": "string", "description": "可选；修改用户名（至少 3 位）" }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "修改成功",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Status200Message" }
              }
            }
          },
          "400": {
            "description": "参数错误（旧密码错误、新密码太短等）",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DetailResponse" }
              }
            }
          },
          "401": {
            "description": "未认证",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DetailResponse" }
              }
            }
          }
        },
        "security": [{ "jwtAuth": [] }]
      }
    },
    "/web/auth/verify-password": {
      "post": {
        "summary": "验证密码",
        "description": "验证当前用户密码是否正确，用于敏感操作前的身份确认。",
        "operationId": "verifyPassword",
        "tags": ["web-auth"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["password"],
                "properties": {
                  "password": { "type": "string" }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "验证通过",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Status200Message" }
              }
            }
          },
          "400": {
            "description": "参数错误",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DetailResponse" }
              }
            }
          },
          "401": {
            "description": "密码错误或未认证",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DetailResponse" }
              }
            }
          }
        },
        "security": [{ "jwtAuth": [] }]
      }
    },
...
    "/web/users/{id}": {
      "put": {
        "summary": "更新用户",
        "operationId": "updateUser",
        "tags": ["web-users"],
        "parameters": [
          {
            "name": "id", "in": "path", "description": "用户 ID", "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": { "type": "string" },
                  "password": { "type": "string", "minLength": 6 },
                  "role": { "type": "string", "enum": ["admin", "school_w", "grade_w", "class_w", "readonly"] },
                  "scope": { "type": "string" },
                  "must_change_pwd": { "type": "boolean" },
                  "must_change_username": { "type": "boolean" }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "更新成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["status", "user"],
                  "properties": {
                    "status": { "type": "integer", "example": 200 },
                    "user": { "$ref": "#/components/schemas/UserInfo" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "参数错误",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DetailResponse" }
              }
            }
          },
          "401": { "description": "未认证或无 admin 权限" },
          "404": {
            "description": "用户不存在",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DetailResponse" }
              }
            }
          }
        },
        "security": [{ "jwtAuth": [] }]
      },
      "delete": {
        "summary": "删除用户",
        "operationId": "deleteUser",
        "tags": ["web-users"],
        "parameters": [
          {
            "name": "id", "in": "path", "description": "用户 ID", "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": {
            "description": "删除成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["status", "deleted"],
                  "properties": {
                    "status": { "type": "integer", "example": 200 },
                    "deleted": { "type": "integer", "format": "int64" }
                  }
                }
              }
            }
          },
          "401": { "description": "未认证或无 admin 权限" },
          "404": {
            "description": "用户不存在",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DetailResponse" }
              }
            }
          }
        },
        "security": [{ "jwtAuth": [] }]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CompensationSingleResponse": {
        "type": "object",
        "properties": {
          "date": { "type": "string", "format": "date", "description": "查询的日期" },
          "compensation": {
            "type": "string",
            "format": "date",
            "nullable": true,
            "description": "对应的调休日期；无调休时为 null"
          }
        }
      },
      "CompensationYearResponse": {
        "type": "object",
        "properties": {
          "year": { "type": "integer" },
          "pairs": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "holiday": { "type": "string", "format": "date", "description": "节假日日期" },
                "workday": { "type": "string", "format": "date", "description": "补班工作日" }
              }
            }
          }
        }
      },
      "Period": {
        "type": "object",
        "properties": {
          "no": { "type": "integer", "description": "课节序号" },
          "subject": { "type": "string", "description": "科目简称" }
        }
      },
...
      "DetailResponse": {
        "type": "object",
        "properties": {
          "detail": { "type": "string" }
        }
      },
...
      "Status200Message": {
        "type": "object",
        "required": ["status", "message"],
        "properties": {
          "status": { "type": "integer", "example": 200 },
          "message": { "type": "string" }
        }
      },
      "UserInfo": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "username": { "type": "string" },
          "role": { "type": "string", "enum": ["admin", "school_w", "grade_w", "class_w", "readonly"] },
          "scope": { "type": "string" }
        }
      },
...
    },
    "securitySchemes": {
      "basicAuth": {
        "type": "http",
        "scheme": "basic",
        "description": "（已废弃，仅供旧版客户端兼容）用户名：`AstraSchedule` 或 `ElectronClassSchedule`，密码：配置文件 `secret.token`"
      },
      "jwtAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Token，通过 POST /web/auth/login 获取。写入接口还需在 X-Verify-Password 头中提供密码。"
      }
    }
  },
  "servers": [
    {
      "url": "https://{hostvar1}.getastra.cn"
    }
  ],
  "security": []
}